+++ Impact/nefesh/src/stylesheets/page/AdminPage.css
