+++ Impact/nefesh/src/stylesheets/scrollbar.css
@@ -1,0 +1,30 @@
+html.dark {
+    --theme-scrollbar-thumb: #202225;
+    --theme-scrollbar-track: #2f3136;
+}
+
+html.light {
+    --theme-scrollbar-thumb: #ccc;
+    --theme-scrollbar-track: #f2f2f2;
+}
+
+::-webkit-scrollbar {
+    height: 16px;
+    width: 16px;
+}
+
+::-webkit-scrollbar-thumb {
+    background-color: var(--theme-scrollbar-thumb);
+    border: 4px solid transparent;
+    border-radius: 8px;
+    min-height: 40px;
+    background-clip: padding-box;
+}
+
+::-webkit-scrollbar-track {
+    background-color: var(--theme-scrollbar-track);
+    border: 4px solid transparent;
+    border-radius: 8px;
+    margin-bottom: 8px;
+    background-clip: padding-box;
+}
