+++ Impact/nefesh/src/views/pages/Login.js
@@ -1,11 +1,11 @@
-import { db } from '../../../functions/firebase/config';
+import { db } from '../../functions/firebase/config';
 import React from 'react';
-import { authentication } from '../../../functions/firebase/config';
+import { authentication } from '../../functions/firebase/config';
 import { signInWithPopup, GoogleAuthProvider } from 'firebase/auth';
-import './Login.css';
-import NefishLogo from "../../components/Images/Nefish_B'Nefish_Logo.png";
-import GoogleLogo from '../../components/Images/Google.png';
-import LoginArt from '../../components/Images/signin_img.svg';
+import '../../stylesheets/page/Login.css';
+import NefishLogo from "../../assets/Images/Nefish_B'Nefish_Logo.png";
+import GoogleLogo from '../../assets/Images/Google.png';
+import LoginArt from '../../assets/Images/signin_img.svg';
 
 function Login() {
 	const SignIn = (ev) => {
