+++ Impact/nefesh/src/views/pages/MainPage.jsx
@@ -1,6 +1,6 @@
-import "./MainPage.css";
+import "../../stylesheets/page/MainPage.css";
 import { useEffect, useState } from "react"
-import { db } from '../../../functions/firebase/config';
+import { db } from '../../functions/firebase/config';
 import { getDatabase, ref, onValue, query } from "firebase/database";
 import { collection,  orderBy, onSnapshot, getDocs, where, getFirestore} from 'firebase/firestore';
