+++ Impact/nefesh/src/views/pages/ProfilePage1.jsx
@@ -1,6 +1,6 @@
-import './ProfilePage.css';
+import '../../stylesheets/page/ProfilePage.css';
 import React, { useEffect, useState } from 'react';
-import { db } from '../../../functions/firebase/config';
+import { db } from '../../functions/firebase/config';
 import { doc, getDoc, onSnapshot } from 'firebase/firestore';
 
 let userDocRef
