+++ Impact/nefesh/src/views/templates/AdminPagePopUp.jsx
@@ -1,9 +1,9 @@
 import React, { useEffect, useState } from 'react';
-import './AdminPagePopUp.css';
-import { db } from '../../../functions/firebase/config';
+import '../../stylesheets/page/AdminPagePopUp.css';
+import { db } from '../../functions/firebase/config';
 import { collection, Firestore } from '@firebase/firestore';
 import { doc, getDoc, getDocs, updateDoc, where } from 'firebase/firestore';
-import { userIDAdm } from '../../pages/AdminPage/AdminPage';
+import { userIDAdm } from '../pages/AdminPage';
 import { query } from 'firebase/database';
 
 let currentID;
