+++ Impact/nefesh/src/views/templates/ArticleCreation.jsx
@@ -1,7 +1,7 @@
 import React, { useState} from 'react';
-import "./ArticleCreation.css"
+import "../../stylesheets/page/ArticleCreation.css"
 import { collection, addDoc, arrayRemove } from "firebase/firestore"
-import { db } from "../../../functions/firebase/config"
+import { db } from "../../functions/firebase/config"
 let i = 0
 let statesSumbitted = { views: 0, startTime: "", endTime: "" }
